Final fix and polish on 2U backplane for OreSat1.
Found a bunch of small little details to fix - somehow CAN lines were down to 0.2 mm, fixed to 0.3 mm. Also expanded some of
the lines to 0.6 mm for higher power lines. Added more vias for VBUS on the debug connector, and finally added 3D models
to all of the connectors so the backplane fully renders now!